requirements.txt
scikit-learn
scikit-image
datasets
opencv-python
numpy
joblib
pillow
matplotlib